Fix bugs of first execution
The first issue is that part of click.get_app_dir path can be missing,
so using makedirs instead of mkdir will avoid errors about missing
parent directories in the path.

On the first run, the config file does not exist and as a result execute
the code to generate it, but the resulting private_key has an invalid
type, so it crashes later in the code with a TypeError message.